Code/requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.27.0
bitsandbytes>=0.42.0
peft>=0.10.0
trl>=0.8.0
outlines>=1.2.9
hydra-core>=1.3.0
pydantic>=2.0.0
wandb>=0.16.0
scikit-learn>=1.4.0
pandas>=2.0.0
tabulate>=0.9.0
pyyaml>=6.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
